+++ js/directives/genericDirectives.js
@@ -14,8 +14,8 @@
 			link: function (scope, element, attrs) {
 				element.on('keydown', function (e) {
 					if ( !!scope.mappings && !!scope.mappings[e.which] ){
-						scope.mappings[e.which]();
-						e.preventDefault();
+						scope.mappings[e.which](e);
+						// e.preventDefault();
 					}
 				});
 			}