+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Major refactoring of the raised adjective, based upon material design guidelines. The `lui_theme_shadow` function thus disappears, and the `.lui.raised` class should not be @extended any longer.
 - Major refactoring of fields and inputs: material style, no more addon support
 - Search input is replaced with a simple `searchable` class to be added to the `div.lui.input`
+- changed the `luid-keydown` directive, it does not automatically call `e.preventDefault()`
 
 ### New features
 - Default border-radius is now 2px instead of 3px